requirements.txt
pyautogui
pillow
watchdog
pytesseract
opencv-python
